assignment_2/requirements.txt
Flask==3.0.0
requests==2.*
finnhub-python